test: Use is_relation_joined to ensure status
