+++ tests/integration/ha_tests/helpers.py
@@ -36,6 +36,7 @@ from ..helpers import (
     get_password,
     mongodb_uri,
     primary_host,
+    is_relation_joined,
 )
 
 METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
@@ -225,21 +226,6 @@ async def deploy_and_scale_application(ops_test: OpsTest) -> str:
         )
 
     return APPLICATION_DEFAULT_APP_NAME
-
-
-def is_relation_joined(ops_test: OpsTest, endpoint_one: str, endpoint_two: str) -> bool:
-    """Check if a relation is joined.
-
-    Args:
-        ops_test: The ops test object passed into every test case
-        endpoint_one: The first endpoint of the relation
-        endpoint_two: The second endpoint of the relation
-    """
-    for rel in ops_test.model.relations:
-        endpoints = [endpoint.name for endpoint in rel.endpoints]
-        if endpoint_one in endpoints and endpoint_two in endpoints:
-            return True
-    return False
 
 
 async def get_process_pid(
